rename lines -> line
