+++ src/lib.rs
@@ -1,13 +1,13 @@
 mod beat_number;
+mod line;
 mod line_launcher;
-mod lines;
 mod message;
 mod midi_clock_tracker;
 mod result;
 
 pub use beat_number::BeatNumber;
+pub use line::Line;
 pub use line_launcher::LineLauncher;
-pub use lines::Line;
 pub use message::Message;
 pub use midi_clock_tracker::MidiClockTracker;
 pub use result::Result;
